Group together helpers in notification_helpers.dart and remove unnecessary files
