+++ lib/providers/notification_provider.dart
@@ -3,9 +3,8 @@ import 'dart:async';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:news_tracker/model/tracked_term.dart';
-import 'package:news_tracker/utils/notifications/initialize_notifications.dart';
 
-import '../utils/notifications/pending_notifications.dart';
+import '../utils/notifications/notification_helpers.dart';
 import '../utils/notifications/schedule_notification.dart';
 
 class NotificationNotifier
