--- lib/utils/notifications/pending_notifications.dart
@@ -1,25 +1,0 @@
-import 'dart:async';
-
-import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-
-import 'initialize_notifications.dart';
-
-Future<List<PendingNotificationRequest>> getPendingNotifications() async {
-  return await notificationsPlugin.pendingNotificationRequests();
-}
-
-Future<PendingNotificationRequest?> getNotificationById(int id) async {
-  final pending = await getPendingNotifications();
-  try {
-    return pending.firstWhere((p) => p.id == id);
-  } catch (_) {
-    return null;
-  }
-}
-
-Future<void> cancelAllNotifications(
-  FlutterLocalNotificationsPlugin? plugin,
-) async {
-  final _plugin = plugin ?? notificationsPlugin;
-  await _plugin.cancelAll();
-}
